+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- ⚡️(backend) hash application secrets with SHA-256
+
+### Fixed
+
+- 🔒️(backend) prevent editing client id and secret in Django admin
+
 ## [1.32.1] - 2026-09-25
 
 ### Fixed